rename: 로또 인터페이스 유효성 검사 관련 파일 뭉침
