+++ src/validators/lottoInterface/checkIsExistInString.js
